Ai/Phone_Ai/requirements.txt
fastapi
uvicorn
transformers
torch
Pillow
faiss-cpu
sentence-transformers
deep-translator
langdetect
jinja2
